feat: Refactor remote files loading logic and move to dedicated route
