+++ packages/insomnia/src/ui/routes/commands.tsx
@@ -21,7 +21,6 @@ import type { RequestGroup } from '../../models/request-group';
 import type { WebSocketRequest } from '../../models/websocket-request';
 import { scopeToActivity, type Workspace } from '../../models/workspace';
 import { invariant } from '../../utils/invariant';
-import { insomniaFetch } from '../insomniaFetch';
 
 export interface CommandItem<TItem> {
   id: string;
@@ -305,81 +304,4 @@ export const loader: LoaderFunction = async args => {
       }),
     },
   };
-};
-
-export interface CommandRemoteItem<TItem> {
-  id: string;
-  url: string;
-  pullUrl: string;
-  name: string;
-  organizationName: string;
-  projectName: string;
-  workspaceName?: string;
-  item: TItem;
-}
-
-interface RemoteFile {
-  id: string;
-  name: string;
-  projectId: string;
-  teamProjectId: string;
-  organizationId: string;
-}
-
-export interface RemoteFilesLoaderResult {
-  files: CommandRemoteItem<RemoteFile & { teamProjectLocalId: string; scope: 'unsynced' }>[];
-}
-
-export const remoteFilesLoader: LoaderFunction = async (): Promise<RemoteFilesLoaderResult> => {
-  const { id: sessionId, accountId } = await userSession.get();
-
-  if (!sessionId) {
-    return {
-      files: [],
-    };
-  }
-
-  try {
-    const remoteFiles = await insomniaFetch<RemoteFile[]>({
-      method: 'GET',
-      path: '/v1/user/files',
-      sessionId,
-    });
-
-    const allOrganizations = JSON.parse(localStorage.getItem(`${accountId}:organizations`) || '[]') as Organization[];
-
-    const allRemoteFilesOrganizationIds = remoteFiles.map(file => file.organizationId);
-    const allRemoteFilesProjectIds = remoteFiles.map(file => file.teamProjectId);
-
-    const organizations = allOrganizations.filter(org => allRemoteFilesOrganizationIds.includes(org.id));
-
-    const projects = await database.find<Project>(project.type, {
-      remoteId: {
-        $in: allRemoteFilesProjectIds,
-      },
-    });
-
-    const files = remoteFiles.map(file => {
-      const parentProject = projects.find(project => project.remoteId === file.teamProjectId);
-      return {
-        id: file.id,
-        url: `/organization/${file.organizationId}`,
-        pullUrl: parentProject
-          ? `/organization/${file.organizationId}/project/${file.teamProjectId}/remote-collections/pull`
-          : '',
-        name: file.name,
-        item: { ...file, teamProjectLocalId: parentProject?._id || '', scope: 'unsynced' as const },
-        organizationName: organizations.find(org => org.id === file.organizationId)?.display_name || '',
-        projectName: parentProject?.name || '',
-      };
-    });
-
-    return {
-      files,
-    };
-  } catch (err) {
-    return {
-      files: [],
-    };
-  }
 };
